Pretend ' ' for empty views
eslint < 4.10.0 has a bug where it shows its CLI usage info when
we send the empty string via stdin. The solution here is to send
one space instead.

Tested with the 'no-trailing-spaces' rule to not mark as an error
for such cases.

Proper version branching is maybe over-engineering, and left as
exercise for the reader.

Another solution would be to just check the configuration bc empty
views usually cannot have lint errors. This *could* be done via
`eslint --print-config .`