Escape $ in file paths
The `$` character in file paths gets treated as a variable by Sublime Text. And as a result, ESLint fails to lint TS files that are "not included" because the path doesn't match any that tsconfig has included.

For example, `/some/long/file.$path.ts` becomes `/some/long/file..ts`

With this change, the `$` is escaped and therefore the full path stays in tact.